Refactor editor suggestions store into usersuggestionmodel
