+++ library/src/scripts/users/suggestion/IUserSuggestion.ts
@@ -1,0 +1,11 @@
+/**
+ * @author Adam (charrondev) Charron <adam.c@vanillaforums.com>
+ * @copyright 2009-2018 Vanilla Forums Inc.
+ * @license GPL-2.0-only
+ */
+
+import { IUserFragment } from "@library/@types/api";
+
+export interface IUserSuggestion extends IUserFragment {
+    domID: string;
+}
