+++ plugins/rich-editor/src/scripts/@types/store.ts
@@ -6,16 +6,7 @@
  */
 
 import { IStoreState as IBaseStoreState } from "@dashboard/@types/state";
-import MentionTrie from "@rich-editor/state/mention/MentionTrie";
 import { RangeStatic } from "quill/core";
-
-export interface IMentionState {
-    lastSuccessfulUsername: string | null;
-    currentUsername: string | null;
-    usersTrie: MentionTrie;
-    activeSuggestionID: string;
-    activeSuggestionIndex: number;
-}
 
 export interface IEditorInstance {
     currentSelection: RangeStatic | null;
@@ -31,6 +22,5 @@ export interface IEditorInstanceState {
 export interface IStoreState extends IBaseStoreState {
     editor: {
         instances: IEditorInstanceState;
-        mentions: IMentionState;
     };
 }
