+++ plugins/rich-editor/src/scripts/state/editorReducer.ts
@@ -4,12 +4,11 @@
  * @license GPL-2.0-only
  */
 
-import mentionReducer from "@rich-editor/state/mention/mentionReducer";
 import instanceReducer from "@rich-editor/state/instance/instanceReducer";
 import { combineReducers } from "redux";
 
 const editorReducer = combineReducers({
-    mentions: mentionReducer,
+    mentions: () => {},
     instances: instanceReducer,
 });
